[refactor/#18] delete unused files
